Move services to app cqrs module
