+++ api/src/app.module.ts
@@ -40,7 +40,8 @@ import { TasksService } from 'modules/tasks/tasks.service';
 import { NotificationsModule } from 'modules/notifications/notifications.module';
 import { ReportsModule } from 'modules/reports/reports.module';
 import { EudrModule } from 'modules/eudr-alerts/eudr.module';
-import { WebSocketsModule } from './modules/notifications/websockets/websockets.module';
+import { WebSocketsModule } from 'modules/notifications/websockets/websockets.module';
+import { AppCqrsModule } from 'modules/cqrs/app-cqrs.module';
 
 const queueConfig: any = config.get('queue');
 
@@ -88,6 +89,7 @@ const queueConfig: any = config.get('queue');
     ReportsModule,
     WebSocketsModule,
     EudrModule,
+    AppCqrsModule,
   ],
   providers: [
     {
