+++ api/src/modules/cqrs/import-data/import-progress.event.ts
@@ -8,6 +8,7 @@ export class ImportProgressUpdateEvent {
 }
 
 export enum IMPORT_PROGRESS_STATUS {
+  VALIDATING_DATA = 'VALIDATING_DATA',
   IMPORTING_DATA = 'IMPORTING_DATA',
   GEOCODING = 'GEOCODING',
   CALCULATING_IMPACT = 'CALCULATING_IMPACT',
